requirements.txt
python-telegram-bot
selenium~=3.141.0
pyperclip